+++ packages/babel-preset-expo/index.js
@@ -3,7 +3,7 @@ const lazyImportsBlacklist = require('./lazy-imports-blacklist');
 let hasWarnedJsxRename = false;
 
 module.exports = function (api, options = {}) {
-  const { web = {}, native = {} } = options;
+  const { web = {}, native = {}, reanimated } = options;
 
   const bundler = api.caller(getBundler);
   const isWebpack = bundler === 'webpack';
@@ -112,6 +112,10 @@ module.exports = function (api, options = {}) {
       platform === 'web' && [require.resolve('babel-plugin-react-native-web')],
       isWebpack && platform !== 'web' && [require.resolve('./plugins/disable-ambiguous-requires')],
       require.resolve('@babel/plugin-proposal-export-namespace-from'),
+
+      // Automatically add `react-native-reanimated/plugin` when the package is installed.
+      hasModule('react-native-reanimated') &&
+        reanimated !== false && [require.resolve('react-native-reanimated/plugin')],
     ].filter(Boolean),
   };
 };